Move program strings to PROGMEM
The user interface strings have been moved to PROGMEM to a reduce RAM
usage. The lcd-progmem.h file defines PSTR2 but it is never used, as
using it in the array definition throws the 'statement-expressions are
not allowed outside functions nor in template-argument lists' error.
Instead of this, I declared the array with the PROGMEM attribute. This
required me to do a few changes in lcd-progmem.h, the original have been
left in comment.